+++ AS1 - Bouncing Balls - Visual Studio Project/AS1 - Bouncing Balls/Registry.h
@@ -11,7 +11,7 @@ using namespace std;
 namespace Registry {
 	extern int window_size[2];
 	extern string game_end_state;
-	extern float ball_radius;
+	extern int ball_radius;
 	extern sf::Music pop_sounds[15];
 	extern sf::Music cannon_fire_sound;
 	extern float ball_tesselation_coefficient;
